CTCP-5565: Removing dependency on /does-ie028-exist-for-lrn endpoint. This will be deprecated.
