--- it/test/connectors/PostTransitionSubmissionConnectorSpec.scala
@@ -1,204 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package connectors
-
-import com.github.tomakehurst.wiremock.client.WireMock._
-import config.{PostTransitionModule, TransitionModule}
-import helpers.{ItSpecBase, WireMockServerHandler}
-import models.{DepartureMessage, DepartureMessages}
-import org.scalacheck.Gen
-import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
-import play.api.inject.guice.GuiceApplicationBuilder
-import play.api.libs.json.JsNumber
-import play.api.test.Helpers._
-import uk.gov.hmrc.http.HttpResponse
-
-class PostTransitionSubmissionConnectorSpec extends ItSpecBase with WireMockServerHandler with ScalaCheckPropertyChecks {
-
-  override def guiceApplicationBuilder(): GuiceApplicationBuilder =
-    super
-      .guiceApplicationBuilder()
-      .configure(conf = "microservice.services.manage-transit-movements-departure-cache.port" -> server.port())
-      .disable[TransitionModule]
-      .bindings(new PostTransitionModule)
-
-  private lazy val connector: SubmissionConnector = app.injector.instanceOf[SubmissionConnector]
-
-  "SubmissionConnector" - {
-
-    "post" - {
-
-      val url = s"/manage-transit-movements-departure-cache/declaration/submit"
-
-      "must return true when status is Ok" in {
-        server.stubFor(
-          post(urlEqualTo(url))
-            .withHeader("Accept", equalTo("application/vnd.hmrc.final+json"))
-            .willReturn(aResponse().withStatus(OK))
-        )
-
-        val result: HttpResponse = await(connector.post(lrn.value))
-
-        result.status mustBe OK
-      }
-
-      "return false for 4xx response" in {
-        val status = Gen.choose(400: Int, 499: Int).sample.value
-
-        server.stubFor(
-          post(urlEqualTo(url))
-            .withHeader("Accept", equalTo("application/vnd.hmrc.final+json"))
-            .willReturn(aResponse().withStatus(status))
-        )
-
-        val result: HttpResponse = await(connector.post(lrn.value))
-
-        result.status mustBe status
-      }
-
-      "return false for 5xx response" in {
-        val status = Gen.choose(500: Int, 599: Int).sample.value
-
-        server.stubFor(
-          post(urlEqualTo(url))
-            .withHeader("Accept", equalTo("application/vnd.hmrc.final+json"))
-            .willReturn(aResponse().withStatus(status))
-        )
-
-        val result: HttpResponse = await(connector.post(lrn.value))
-
-        result.status mustBe status
-      }
-    }
-
-    "postAmendment" - {
-
-      val url = s"/manage-transit-movements-departure-cache/declaration/submit-amendment"
-
-      "must return true when status is Ok" in {
-        server.stubFor(
-          post(urlEqualTo(url))
-            .withHeader("Accept", equalTo("application/vnd.hmrc.final+json"))
-            .willReturn(aResponse().withStatus(OK))
-        )
-
-        val result: HttpResponse = await(connector.postAmendment(lrn.value))
-
-        result.status mustBe OK
-      }
-
-      "return false for 4xx response" in {
-        val status = Gen.choose(400: Int, 499: Int).sample.value
-
-        server.stubFor(
-          post(urlEqualTo(url))
-            .withHeader("Accept", equalTo("application/vnd.hmrc.final+json"))
-            .willReturn(aResponse().withStatus(status))
-        )
-
-        val result: HttpResponse = await(connector.postAmendment(lrn.value))
-
-        result.status mustBe status
-      }
-
-      "return false for 5xx response" in {
-        val status = Gen.choose(500: Int, 599: Int).sample.value
-
-        server.stubFor(
-          post(urlEqualTo(url))
-            .withHeader("Accept", equalTo("application/vnd.hmrc.final+json"))
-            .willReturn(aResponse().withStatus(status))
-        )
-
-        val result: HttpResponse = await(connector.postAmendment(lrn.value))
-
-        result.status mustBe status
-      }
-    }
-
-    "getExpiryInDays" - {
-
-      val url = s"/manage-transit-movements-departure-cache/user-answers/expiry/$lrn"
-
-      "must return expiry in days when status is Ok" in {
-        server.stubFor(
-          get(urlEqualTo(url))
-            .willReturn(okJson(JsNumber(30).toString()))
-        )
-
-        val result: Long = await(connector.getExpiryInDays(lrn.value))
-
-        result mustBe 30
-      }
-    }
-
-    "getMessages" - {
-
-      val url = s"/manage-transit-movements-departure-cache/messages/$lrn"
-
-      val json =
-        """
-          |{
-          |  "messages" : [
-          |    {
-          |      "type" : "IE015"
-          |    },
-          |    {
-          |      "type" : "IE928"
-          |    },
-          |    {
-          |      "type" : "IE013"
-          |    }
-          |  ]
-          |}
-          |""".stripMargin
-
-      "must return messages when status is Ok" in {
-        server.stubFor(
-          get(urlEqualTo(url))
-            .willReturn(okJson(json))
-        )
-
-        val result: DepartureMessages = await(connector.getMessages(lrn))
-
-        result mustBe DepartureMessages(
-          Seq(
-            DepartureMessage("IE015"),
-            DepartureMessage("IE928"),
-            DepartureMessage("IE013")
-          )
-        )
-      }
-
-      "must return empty list when status is NoContent or NotFound" in {
-        forAll(Gen.oneOf(204: Int, 404: Int)) {
-          status =>
-            server.stubFor(
-              get(urlEqualTo(url))
-                .willReturn(aResponse.withStatus(status))
-            )
-
-            val result: DepartureMessages = await(connector.getMessages(lrn))
-
-            result mustBe DepartureMessages(
-              Seq.empty[DepartureMessage]
-            )
-        }
-      }
-    }
-  }
-}
